Add unchecked accesses to class Tensor.
In internal functions, we often have loops over the elements of a tensor for
which we know exactly what indices we use. In these cases, there is not much
of a point to check the validity of these indices every time. Rather, just
use the UncheckedIndex class to indicate that we don't need to check indices
again and again.