Remove the DEAL_II_LIBLAPACK_NOQUERYMODE option.
Some LAPACK functions calculate the optimal work array size that they
need. This option existed to allow compatibility with older versions
that did not support this functionality. This option was never migrated
to CMake and has probably not been necessary in a long time: a comment
refers to functionality missing in PETSc 2.3, which was released in
2005.