Implement parallel MGTransferPrebuilt for parallel::distributed::Vector
This commit includes a refactoring of MGTransferPrebuilt into a base class
MGLevelGlobalTransfer that contains the copy operation between the multigrid
hierarchy and the global degrees of freedom, including a specialization for
parallel::distributed::Vector that takes care of the particular behavior in
this vector type. It also allows us to use optimized access in terms of local
MPI indices.