doxygen: do not export duplicate functions
Since we started handing .cc files to doxygen, every explicit template
declaration of a function inside a .cc would generate another line for
that function in the class documentation (not clickable and before the
correct ones). We fix this by hiding undocumented functions. A quick
test shows, that even with this change, undocumented functions inside
classes defined in the .h will still be exported, because things like
@name groups and @ingroup count as "documentation".