Move some internal functions.
FESystem has a number of functions that are called in the member initializer list
of the constructors. These are all static functions, of course, and don't access
any member variables.

This patch moves them ouf of the class altogether. This has the advantage that
it reduces the size of the .h file, and that it reduces the number of functions
exported by the shared libraries. Rather, these functions may as well live
inside an anonymous namespace in the .cc file.

There are no functional changes, nor anything that would be visible to the user.