Update step-44 to be dimension independent.
By making the appropriate changes to the material class, grid generation
and boundary conditions, the quasi-incompressible indentation problem
can be made dimension independent. The traction is now applied on the +Y
face instead of the +Z face so that the 2-d and 3-d problems can be
post-processed in a similar manner. The 2-d results are qualitatively
similar to that of the 3-d case, but due to the more constrained nature
of the problem (planar motion) the amount of compression achieved is
less than that of the 3-d case (while the laterial extension increases).