Get rid of IterativeInverse in step-20.
IterativeInverse was marked as deprecated in April 2015 in favor of
LinearOperator. This puts this step in a strange place; we should not
use deprecated features in tutorials, but at the same time we cannot use
its replacement, which requires C++11 support. To resolve this, this
step just provides its own replacement for a special case of
IterativeInverse.