Adjust the output of a test to recent changes.
This adjusts a single number, which was previously output as 3.37 and is
now output as 3.38. The correct value is likely 3.375, so will be rounded to
one or the other on the whim of the compiler and the phase of the moon.

The correct approach to fixing this is likely to just output one extra digit
in the .output file. However, this increases the size of the output file by 1.4 MB
and that seems silly, so I will perpetuate the fragile behavior by simply changing
the one number and kicking the can for a proper approach down the road.