confusing set_all_manifold* errors if empty
The error messages for set_all_manifold_ids* are confusing if the mesh
is empty:

An error occurred in line <10760> of file </ssd/deal-
git/source/grid/tria.cc> in function
    dealii::Triangulation<dim, spacedim>::raw_quad_iterator
dealii::Triangulation<dim, spacedim>::begin_raw_quad(unsigned int) const
[with int dim = 2, int spacedim = 2, dealii::Triangulation<dim,
spacedim>::raw_quad_iterator =
dealii::TriaRawIterator<dealii::CellAccessor<2, 2> >]
The violated condition was:
    level<n_global_levels() || level<levels.size()
The name and call sequence of the exception was:
    ExcInvalidLevel(level)
Additional Information:
The given level 0 is not in the valid range!

Fix that.