Update a copy-paste error for DoFTools::extract_boundary_dofs().
There are two functions of the same name with the same documentation, but in one
the argument is a std::vector<bool> and in the other it is an IndexSet. The former's
documentation was copied from the latter's and referred to the wrong data type.

While there, also mark arguments as [in] or [out].