Use index sets to simplify some assertions.
GCC 6.1 now warns (-Waddress) that (since the dynamic cast of an object
back to its own type always succeeds) the address of an object is always
nonzero. Fortunately these dynamic casts are not needed anymore due to
better support for generic programming in the library.

Since local_size is not available for scalar vectors, perform (stricter)
checks with index sets instead.

Regardless of the new GCC warning, these assertions are useful for all
distributed vectors (and may be run now for sequential vectors too) and
dof handler type, so they should always be checked.