fix FEFieldFunction in parallel
This fixes ExcPointNotAvailableHere() thrown from several places inside
FEFieldFunction, because we assert that the cell is locally owned. This
is not true, because we can happen to look at a ghost cell. Of course it
is totally fine to evaluate function values on ghost cells if we have a
ghosted solution vector.
Triggering all code paths in the test is somewhat tricky, because we
need find_active_cell_around_point to return a ghost cell.