don't use ninja for testsuite
Do not pass the generator to the testsuite so that we use make instead
of ninja for the test projects. This is because calling ninja several
times in parallel for the same project will break the configuration.

This shouldn't have a performance penalty because the individual targets
are relatively simple and don't depend on more than one file.