Initialize a member variable.
This is not strictly necessary, because the variable is later initialized in
constructors of derived classes, but the patch avoids copying uninitialized
variables when copying around objects of type FiniteElementData. (There are
cases when we pass around these objects, without actually using derived
classes; but in those cases, the flag appears to be unused.