Move 'cached_primitivity' and 'is_primitive()'.
These fields were part of FiniteElementData, but this class had no way of computing
the information during construction, and it needed to be set at a later time by
the derived classes' construction time. This is silly. Move the field and function
to class FiniteElement, where we know how to compute the information, and where we
already store similar information.