Correct test outputs.
On my desktop, I have not installed numdiff yet, and these tests fail because
there are either differences in whitespace, or there is a trailing colon on
a line. Neither seem to bother numdiff, but they bother regular diff. Fix this
by checking in corrected versions of these output files.