Tighten up error messages.
It has long annoyed me that when we print exception messages that were created
via ExcMessage, that the information is really duplicative. Example:

--------------------------------------------------------
An error occurred in line <710> of file <.../step-6.cc> in function
    int main()
The violated condition was:
    false
The name and call sequence of the exception was:
    ExcMessage ("abc" "def")
Additional Information:
abcdef

Stacktrace:
-----------
#0  ./step-6: main
--------------------------------------------------------

Here, the text that comes after 'The name and call sequence...'
is duplicative with what comes after 'Additional Information:'.

This is specifically the case for ExcMessage errors because
there *everything* that constructs the 'Additional information'
will actually be also under 'The name and...' part.

This patch cleans this up by omitting the first of these two
parts if the error message was created via ExcMessage.