Generalize the data structure used to represent an edge set.
Specifically, the class is used to represent the set of edges by which the parallel
set can grow in each iteration. In 2d, this set can only have zero, one, or two
elements, and we can use a corresponding, O(1) data structure to store them.

In 3d, the set can be of arbitrary size and we represent it as a std::set.

Creating generic data structures also allows to write the algorithm that uses it
in a more concise way using iterators.