Fix GridIn::read_vtk().
The function reads a section of VTK files of the form

CELL_DATA 4
SCALARS MaterialID double 1
LOOKUP_TABLE default
1.000000
2.000000
3.000000
4.000000

but internally (despite the fact that the field is declared
as 'double' above) reads them as integers. This leads to
wrong results if they are given as floating point numbers,
as in the case above.

While there also add an assertion an clean things up
slightly.