Let DataOut::build_one_patch use a member variable.
Currently, it takes a reference to a variable as argument, but this variable
is a member variable of the class to which it has access. That's duplicative.
To make things even more confusing, the local variable and the member
are named the same.

Fix this by just not taking the reference as argument, and directly
accessing the member of the class.