Rename a function argument.
DataOutRotation::build_one_patch took a variable named 'patches',
but unlike in #3445, it is not a reference to the class's member
variable of the same name. Rather, it is an array this function
needs to fill in, and nobody else sees it. Consequently,
rename it 'my_patches'.

While there, also update the documentation, which is no longer
in agreement with the implementation.