Fix an issue with an MPI implementation.
I seem to be running into an MPI implementation problem whereby calling
MPI_Allreduce returns garbage in the output buffer if (i) we are
working on 'unsigned long long' data types, and (ii) we call the
allreduce function on only one processor. Initializing the output
buffer with something seems to help the problem, but the cheaper
alternative is to just not call MPI_Allreduce if we are on one processor
only.