Avoid a memory leak.
In fact, the test really tries very hard to create a memory leak, and
even verifies that it happens. But this then leads to downstream heartbreak
if you run a memory checker because you get pulled over for it.

Fix the issue by cleaning up the memory leak after we have verified that
the code that ran before really did leak the memory, by keeping a pointer
to the leaked object after all, and deleting it at the end of it all.