Move template function DoFHandler::n_boundary_dofs().
This function is currently in a .cc file, and the .inst.in file actually tries to
instantiate it for all template arguments. Nonetheless, as #3599 shows, it is not
instantiated appropriately.

Rather than try to understand the exact cause, the easier solution is to just
move the function to the .h file. It turns out to have a lot of code duplicated
from the other function of same name, so shrink it down to the minimum and
simply defer to the non-templated member function.