Fix and enable compression of hp DoF indices.
A previous patch, a long time ago, implemented a scheme to compress the data structure
that stores the DoF indices in the hp case. For reasons no longer clear, the functions
that do the compression and decompression just returned immediately, without actually
doing anything. Maybe not surprisingly, the algorithm -- not executed after all -- did
not work.

This patch fixes the algorithm at various locations and enables the compression.