Refactor identity_operator and null_operator for non-trivial payloads
When performing compound operations with LinearOperators, it is
necessary for the Payload type of all operations to be the same.
identity_operator did not take in a operator, so we used to define a
variant of it in the TrilinosWrappers namespace in order to set up the
underlying Payload. This lead to some compiler confusion as it could not
deduce which version of the function to call unless the namespace was
used to explicitly remove this ambiguity.

We've now removed the second identity_operator in TrilinosWrappers, and
have instead defined an alternative identity_operator that takes in an
exemplar LinearOperator. This version of the function will automatically
extract the range of operations from the exemplar, deduce the Payload
type AND initialise it correctly (this would not have been done
correctly before). This required the introduction of the inverse_payload
function to the Payload classes.

In anticipation of similar problems with null_operator, this
LinearOperator is also slightly refactored as to ensure that the payload
is correctly configured. This required the introduction of the
null_payload function to the Payload classes.

Fixes #3846