Remove a check on a timer slope.
This commit gets rid of the part of a test that tried to check the
computational complexity of distributing DoFs by fitting a least squares
curve to timing data. This frequently failed and it is not really
needed: the test will time out if we ever go back to a quadratic time
algorithm.