Enable deal.II solver use with Trilinos LinearOperator
By making a distinction between an `inverse_payload` created for
Trilinos solvers/preconditioners and all other solver/preconditioner
types, we allow a `TrilinosPayload` to be created when the
solver/preconditioner is incompatible with `Epetra_Multivector`. In
turn, it is now possible to use the deal.II solvers (e.g.
`dealii::SolverCG<TrilinosWrappers::MPI::Vector>`) with `Trilinos`
`LinearOperator`s as such a call no longer results in a compiler error.