CMake: Workaround: Ensure "mpiuni/mpi.h" is reachable
In case of a sequential build without mpi support and sequential petsc a
dummy mpi.h file installed by petsc must be reachable in from our final
set of include directories.

Normally this is not an issue - petsc exports such information in the
file "petscvariables" that we parse. But some distributions have chosen
to not install this file...

Closes #4157