Add class ImmersedSurfaceQuadrature
Add a class representing an quadrature rule over a surface cutting arbitrary
through the unit cell. In order to map a quadrature rule over a surface from
the unit cell to physical space it is not sufficient to know only points and
scalar weights, since information about the normal is also required. In
addition to points and weights this class also stores the normal vectors.