Split DataOut_DoFData instantiations.
Before this commit GCC needed 2.3 GB, 65s in release mode and 3.6 GB, 79s in
debug mode to compile this file. After splitting the file we need 43s/32s and
1.6 GB/1.3 GB in release mode and 52s/34s and 2.1 GB/2.0 GB in debug mode. This
is substantial because this was (previously) the most expensive file to
compile.

This also adds a few missing instantiations to the nonzero codimension case.