Change some indentation oddities.
The majority of these are related to space before comma. I did all of
this by first running astyle 3.0.1 to see what it changes, and then
running astyle 2.04 as usual again. 3.0.1 fixes up a few oddities
that 2.04 doesn't undo, and most of these are probably worthwhile
keeping.

There are some corner cases where we had purposefully aligned commas
with leading spaces, and one could consider not making this change,
or updating it by hand. But I don't feel strongly enough about
this to really go to lengths here.